+++ src/icons/HarnessIcons.ts
@@ -4,6 +4,7 @@ import Edit from './Edit.svg'
 import Inline from './Inline.svg'
 import Options from './Options.svg'
 import Activity from './activity.svg'
+import Adminrole from './adminRole.svg'
 import AppAwsCodeDeploy from './app-aws-code-deploy.svg'
 import AppAwsLambda from './app-aws-lambda.svg'
 import AppKubernetes from './app-kubernetes.svg'
@@ -84,6 +85,7 @@ import CoverageStatusError from './coverage-status-error.svg'
 import CoverageStatusSuccess from './coverage-status-success.svg'
 import CsHover from './cs-hover.svg'
 import CustomService from './custom-service.svg'
+import Customrole from './customRole.svg'
 import CvHover from './cv-hover.svg'
 import CvMain from './cv-main.svg'
 import DashboardSelected from './dashboard-selected.svg'
@@ -246,6 +248,7 @@ import MainView from './main-view.svg'
 import MainWarning from './main-warning.svg'
 import MainWorkflows from './main-workflows.svg'
 import MainZoomOut from './main-zoom-out.svg'
+import Memberrole from './memberRole.svg'
 import MoneyIcon from './money-icon.svg'
 import MultiService from './multi-service.svg'
 import NavAccountAdminHover from './nav-account-admin-hover.svg'
@@ -430,6 +433,7 @@ import TriggerSchedule from './trigger-schedule.svg'
 import Up from './up.svg'
 import Utility from './utility.svg'
 import ViewJson from './view-json.svg'
+import Viewerrole from './viewerRole.svg'
 import Waiting from './waiting.svg'
 import WhiteCluster from './white-cluster.svg'
 import WhiteFullCluster from './white-full-cluster.svg'
@@ -447,6 +451,7 @@ type HarnessIconName =
   | 'Inline'
   | 'Options'
   | 'activity'
+  | 'adminRole'
   | 'app-aws-code-deploy'
   | 'app-aws-lambda'
   | 'app-kubernetes'
@@ -527,6 +532,7 @@ type HarnessIconName =
   | 'coverage-status-success'
   | 'cs-hover'
   | 'custom-service'
+  | 'customRole'
   | 'cv-hover'
   | 'cv-main'
   | 'dashboard-selected'
@@ -689,6 +695,7 @@ type HarnessIconName =
   | 'main-warning'
   | 'main-workflows'
   | 'main-zoom-out'
+  | 'memberRole'
   | 'money-icon'
   | 'multi-service'
   | 'nav-account-admin-hover'
@@ -873,6 +880,7 @@ type HarnessIconName =
   | 'up'
   | 'utility'
   | 'view-json'
+  | 'viewerRole'
   | 'waiting'
   | 'white-cluster'
   | 'white-full-cluster'
@@ -890,6 +898,7 @@ const HarnessIcons: KVO<FunctionComponent<ElementType>> = {
   Inline: Inline,
   Options: Options,
   activity: Activity,
+  adminRole: Adminrole,
   'app-aws-code-deploy': AppAwsCodeDeploy,
   'app-aws-lambda': AppAwsLambda,
   'app-kubernetes': AppKubernetes,
@@ -970,6 +979,7 @@ const HarnessIcons: KVO<FunctionComponent<ElementType>> = {
   'coverage-status-success': CoverageStatusSuccess,
   'cs-hover': CsHover,
   'custom-service': CustomService,
+  customRole: Customrole,
   'cv-hover': CvHover,
   'cv-main': CvMain,
   'dashboard-selected': DashboardSelected,
@@ -1132,6 +1142,7 @@ const HarnessIcons: KVO<FunctionComponent<ElementType>> = {
   'main-warning': MainWarning,
   'main-workflows': MainWorkflows,
   'main-zoom-out': MainZoomOut,
+  memberRole: Memberrole,
   'money-icon': MoneyIcon,
   'multi-service': MultiService,
   'nav-account-admin-hover': NavAccountAdminHover,
@@ -1316,6 +1327,7 @@ const HarnessIcons: KVO<FunctionComponent<ElementType>> = {
   up: Up,
   utility: Utility,
   'view-json': ViewJson,
+  viewerRole: Viewerrole,
   waiting: Waiting,
   'white-cluster': WhiteCluster,
   'white-full-cluster': WhiteFullCluster,